requirements.txt
google-genai>=1.0.0
requests>=2.31.0
flask>=3.0.0
